Add files via upload
This compares the performance of SVM and Naive Bayes models in classifying raisin types (Kecimen and Besni). The notebook includes steps for data preprocessing, feature correlation analysis, and model training. It evaluates the models using confusion matrices and classification reports, providing insights into which model is better suited for this task.